# Bash-it completions performance improvement