@tsiflimagas 2021-12-11
No problem!😄
Damn it, it seems that this implementation doesn't work on Macos. I'm not sure why, but I suspect it could be because of it having BSD grep, instead of GNU one (though I looked into some documentation and it seems that it supports the options I used). Also, build-docs test seems to have failed due to inability to install some python package (I guess it's unrelated to my changes). Anyway, I figured out I could replace external programs with bash features more extensively, which I guess will improve portability and improves speed a bit further. I hope it fixes execution on Macos. It just looks ugly that I repeat variable names like that, so please tell me if you think of something better.